Move test builder from expand test to rust writer tests
