+++ crates/pop-common/src/rust_writer/expand/tests.rs
@@ -1,50 +1,10 @@
 // SPDX-License-Identifier: GPL-3.0
 
 use super::*;
-use crate::rust_writer::helpers;
-use std::{fs, path::PathBuf};
+use crate::rust_writer::tests::test_builder::TestBuilder;
 use syn::parse_str;
 
-struct TestBuilder {
-	test_files: PathBuf,
-	ast: File,
-}
-
-impl Default for TestBuilder {
-	fn default() -> Self {
-		Self {
-			test_files: PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-				.join("src")
-				.join("rust_writer")
-				.join("test_files"),
-			ast: parse_quote! {},
-		}
-	}
-}
-
-macro_rules! add_ast_to_builder{
-    ($([$name: ident, $file: literal $(, $macro_excluded: literal)?]),*) => {
-        $(
-            fn $name(&mut self){
-            self.ast = helpers::preserve_and_parse(
-                fs::read_to_string(self.test_files.join($file))
-                    .expect(concat!{"Error reading file in ", stringify!($name)}),
-                vec![$($macro_excluded)?])
-                .expect(concat!{"Error parsing file in ", stringify!($name)});
-            }
-        )*
-    };
-}
-
 impl TestBuilder {
-	add_ast_to_builder! {
-		[add_basic_pallet_ast, "basic_pallet.rs"],
-		[add_basic_pallet_with_config_preludes_ast, "basic_pallet_with_config_preludes.rs"],
-		[add_outer_config_preludes_ast, "outer_config_preludes.rs"],
-		[add_runtime_using_runtime_macro_ast, "runtime_using_runtime_macro.rs"],
-		[add_runtime_using_construct_runtime_macro_ast, "runtime_using_construct_runtime_macro.rs", "construct_runtime"]
-	}
-
 	fn assert_item_in_config_trait(&self, contains: bool, checked_item: TraitItem) {
 		let mut assert_happened = false;
 		for item in &self.ast.items {
