+++ crates/pop-common/src/rust_writer/tests.rs
@@ -1,0 +1,3 @@
+// SPDX-License-Identifier: GPL-3.0
+
+pub(crate) mod test_builder;
